Squashed commit of the following:
commit 5e2b4c58bdd7f758fa7ab21652c478f9ab6d9498
Author: Ryan Rathsam <ryanrath@buffalo.edu>
Date:   Wed Nov 1 14:53:17 2017 -0400

    Updated ConfigFilesMigration so that it brings roles.json up to snuff

    - ConfigFilesMigration will now ensure that the information required for this
      version of xdmod is present.

commit adaa0aedc8b94479730b36af0b59b589cff2e55e
Author: Ryan Rathsam <ryanrath@buffalo.edu>
Date:   Wed Nov 1 14:10:37 2017 -0400

    Bringing this files formatting up to snuff

commit 76c013519500c3cbc12cbc1b0c160580899e9e9c
Author: Ryan Rathsam <ryanrath@buffalo.edu>
Date:   Wed Nov 1 13:27:54 2017 -0400

    Handle when a user has an institution

    - The previous code was utilizing the old $role_config layout and so when
      updating a user w/ an institution it would bork. This change keeps the logic
      of the old code intact while utilizing the updated role_config ( aka. acls )
      layout.

commit 925817279b6278da677b15a624d64f058a38fe98
Author: Ryan Rathsam <ryanrath@buffalo.edu>
Date:   Wed Nov 1 09:49:59 2017 -0400

    Changing XDUser::setOrganizations to fail fast

    - or at least faster... Now if the system is unable to identify an acl that
      corresponds with the role the caller provided then it will exception out (
      with a hopefully descriptive error message ) so as to prevent the db from
      getting out of sync.

commit 0e12e785887da5a965e7dace532599b1b8749971
Author: Ryan Rathsam <ryanrath@buffalo.edu>
Date:   Wed Nov 1 09:35:33 2017 -0400

    Removing unneeded code from getDisabledMenus

commit 6a2accbdecd073506724e97705d5943069666c15
Author: Ryan Rathsam <ryanrath@buffalo.edu>
Date:   Wed Nov 1 09:15:21 2017 -0400

    Shortening lines to make Travis Happy

commit 36ac6470c891a4f9f88601ce064165ee87f14ac9
Author: Ryan Rathsam <ryanrath@buffalo.edu>
Date:   Wed Nov 1 09:10:20 2017 -0400

    Removing spaces in new 'requires' property.

commit 7926c8259f7b793d65f66a3b45746eafd82433a4
Author: Ryan Rathsam <ryanrath@buffalo.edu>
Date:   Wed Nov 1 09:04:26 2017 -0400

    Adding the 'requires' property to the expected test data.

commit 9e23aaf73cb8f4f64866e8204e315f3b5710981d
Author: Ryan Rathsam <ryanrath@buffalo.edu>
Date:   Tue Oct 31 19:41:35 2017 -0400

    Handling acls that require centers to be provided

    - Actually added the code that handles processing / reporting when an acl
      requires additional information to function properly. In particular, when an
      acl requires a 'provider' ( center / service_provider ).

+++ classes/XDUser.php
@@ -1635,19 +1635,22 @@ SQL
             throw new Exception("This user must be saved prior to calling setOrganization()");
         }
 
+        $acl = Acls::getAclByName($role);
+        if (!isset($acl)) {
+            throw new Exception("Cannot locate an acl for [$role]. Unable to continue setting organizations for user: " . $this->getUsername());
+        }
+
         $role_id = $this->_getRoleID($role);
         if (null === $role_id) {
             throw new Exception("Unable to retrieve id for role: $role");
         }
-        $acl = Acls::getAclByName($role);
-
-        // -------------------------------------------------------
 
         $this->_pdo->execute("DELETE FROM UserRoleParameters " .
             "WHERE user_id=:user_id AND role_id=:role_id AND param_name='provider'", array(
             ':user_id' => $this->_id,
             ':role_id' => $role_id,
         ));
+
         $this->_pdo->execute(
             "DELETE FROM user_acl_group_by_parameters WHERE user_id = :user_id AND acl_id = :acl_id AND group_by_id IN (SELECT gb.group_by_id FROM group_bys gb WHERE gb.name = 'provider')",
             array(
@@ -1655,6 +1658,9 @@ SQL
                 ':acl_id' => $acl->getAclId()
             )
         );
+
+        // -------------------------------------------------------
+
         // =======================================
 
         $active_is_in_set = false;
@@ -1699,8 +1705,10 @@ SQL
                 ':is_primary' => $primary_flag,
                 ':is_active' => $active_flag,
             ));
-            $this->_pdo->execute(
-                <<<SQL
+
+            if (isset($acl)) {
+                $this->_pdo->execute(
+                    <<<SQL
 INSERT INTO user_acl_group_by_parameters (user_id, acl_id, group_by_id, value) 
 SELECT inc.* 
 FROM (
@@ -1719,13 +1727,14 @@ LEFT JOIN user_acl_group_by_parameters cur
   AND cur.value = inc.value
 WHERE cur.user_acl_parameter_id IS NULL;
 SQL
-                ,
-                array(
-                    ':user_id' => $this->_id,
-                    ':acl_id' => $acl->getAclId(),
-                    ':value' => $organization_id
-                )
-            );
+                    ,
+                    array(
+                        ':user_id' => $this->_id,
+                        ':acl_id' => $acl->getAclId(),
+                        ':value' => $organization_id
+                    )
+                );
+            }
         }//foreach
 
         // =======================================
